fixup! refactor(ivy): remove merge mode in favor of direct-mode styling
