+++ packages/core/src/render3/styling/binding_concatenation.ts
